% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_evaluation_plotting_functions.R
\name{create_love_plot}
\alias{create_love_plot}
\title{Creates a love plot}
\usage{
create_love_plot(
  input,
  title = "untitled",
  subtitle = "untitled",
  order_by = "unmatched",
  horizontal = FALSE,
  segment_size = 0.3,
  point_size = 1,
  segment_color = "#b0aba5",
  matched_color = "#30694a",
  prematched_color = "#9c513a",
  background_color = "#F1EEE9",
  axis_text_size = 8,
  axis_text_angle = 45,
  axis_text_hjusts = 1,
  legend_position = "bottom"
)
}
\arguments{
\item{input}{Dataframe created from the create_plot_metadata function.}

\item{title}{A string to be used as the title of the plot.}

\item{subtitle}{A string to be used as the subtitle of the plot.}

\item{order_by}{Defaults to "unmatched". A string indicating how the y-axis should be ordered, options are "unmatched", "matched", or "difference".}

\item{horizontal}{Defaults to FALSE. When TRUE, the plot will be horizontal instead of vertical,}

\item{segment_size}{Defaults to 0.3. The size of the segments connecting treatment and control dot geoms.}

\item{point_size}{Defaults to 1. The side of the point geoms.}

\item{segment_color}{Defaults to "#b0aba5". A string indicating the color of the segments connecting treatment and control dot geoms.}

\item{matched_color}{Defaults to "#30694a". A string indicating the color of the dot geoms representing the matched set value.}

\item{prematched_color}{Defaults to "#9c513a". A string indicating the color of the dot geoms representing the prematched set value.}

\item{background_color}{Defaults to "#f1eee9". A string indicating the color of the plot background}

\item{axis_text_size}{Defaults to 8. An integer indicating the font size along both axis.}

\item{axis_text_angle}{Only applies when horizontal is TRUE, defaults to 45.}

\item{axis_text_hjusts}{Only applies when horizontal is TRUE, defaults to 1.}

\item{legend_position}{Defaults to "bottom". takes any string the legend.position argument excepts in ggplot::theme.}
}
\description{
This function takes in a dataframe created with the create_plot_metadata function
as an input and outputs a loveplot
}
\examples{
# load the MathIt Package and load the included "lalonde" dataset
library("MatchIt")
data("lalonde", package = "MatchIt")

# Create a MatchIt Object using the matchit() function
match <- matchit(treat ~ age + educ + race + married + nodegree + re74 + re75,
                 data = lalonde,
                 distance = "mahalanobis",
                 replace = TRUE)

create_love_plot(match)

# if something other than MatchIt was used to get a matched dataset, you can use a dataframe with the following
# columns as input
# - "covariate" (character)
# - "unmatched_std_mean_diff" (numeric)
# - "matched_std_mean_diff" (numeric)
}
