% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_evaluation_plotting_functions.R
\name{create_plot_metadata}
\alias{create_plot_metadata}
\title{Create a dataframe of pre and post matched standard mean difference between treatment and control groups}
\usage{
create_plot_metadata(match_object)
}
\description{
This function takes in a MatchIt object and returns a dataframe that can used to
 create love plots using the create_love_plot function
}
\details{
@param match_object MatchIt object that contains metadate about pre and post match dataframes
 @export
}
