% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_evaluation_plotting_functions.R
\name{create_tiefighter_plot}
\alias{create_tiefighter_plot}
\title{Creates "tie-fighter" plot}
\usage{
create_tiefighter_plot(
  estimate_objects,
  x_title,
  title,
  subtitle,
  background_color = "#F5E8E4"
)
}
\arguments{
\item{estimate_objects}{One or more glm objects give as a list.}

\item{x_title}{String used as the title for the x-axis.}

\item{title}{String used as the title for the plot.}

\item{subtitle}{String used as the subtitle for the plot.}

\item{background_color}{Defaults to "#f5e8e4". A string indicating the plot background color.}
}
\description{
Takes in one or more glm opjects and turnes creates a "tie-fighter" plot using
the estimates and standard errors
}
