% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_evaluation_plotting_functions.R
\name{create_love_plot}
\alias{create_love_plot}
\title{Creates a love plot}
\usage{
create_love_plot(
  input,
  title,
  subtitle,
  horizontal = FALSE,
  segment_size = 0.3,
  segment_color = "#b0aba5",
  matched_color = "#30694a",
  prematched_color = "#9c513a",
  background_color = "#F1EEE9",
  axis_text_size = 8,
  axis_text_angle = 45,
  axis_text_hjusts = 1,
  legend_position = "bottom"
)
}
\arguments{
\item{input}{Dataframe created from the create_plot_metadata function.}

\item{title}{A string to be used as the title of the plot.}

\item{subtitle}{A string to be used as the subtitle of the plot.}

\item{horizontal}{Defaults to FALSE. When TRUE, the plot will be horizontal instead of vertical,}

\item{segment_size}{Defaults to 0.3. The size of the segments connecting treatment and control dot geoms.}

\item{segment_color}{Defaults to "#b0aba5". A string indicating the color of the segments connecting treatment and control dot geoms.}

\item{matched_color}{Defaults to "#30694a". A string indicating the color of the dot geoms representing the matched set value.}

\item{prematched_color}{Defaults to "#9c513a". A string indicating the color of the dot geoms representing the prematched set value.}

\item{background_color}{Defaults to "#f1eee9". A string indicating the color of the plot background}

\item{axis_text_size}{Defaults to 8. An integer indicating the font size along both axis.}

\item{axis_text_angle}{Only applies when horizontal is TRUE, defaults to 45.}

\item{axis_text_hjusts}{Only applies when horizontal is TRUE, defaults to 1.}

\item{legend_position}{Defaults to "bottom". takes any string the legend.position argument excepts in ggplot::theme.}
}
\description{
This function takes in a dataframe created with the create_plot_metadata function
as an input and outputs a loveplot
}
